backend/requirements.txt
# Core
Django>=5.1.3
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
django-filter>=24.3
django-extensions>=3.2.3

# Authentication & Email
djangorestframework-simplejwt>=5.3.1
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.158.0
django-allauth>=0.57.0  # For OAuth2 support
social-auth-app-django>=5.4.0  # For social authentication
django-oauth-toolkit>=2.3.0  # For OAuth2 provider functionality
django-templated-email>=3.0.1  # For email templates

# Email Backend Options
django-ses>=3.5.2  # Amazon SES (optional for production)
django-smtp-ssl>=1.0  # For secure SMTP
django-anymail>=10.1  # For multiple email backend support

# API and Web Frameworks
FastAPI>=0.115.6
pydantic>=2.10.4
pydantic-core>=2.27.2
anyio>=4.7.0
h11>=0.14.0

# Database and Search
mysql-connector>=2.2.9
mysqlclient>=2.2.6
elasticsearch>=8.17.0
elastic-transport>=8.15.1

# File Handling
Pillow>=11.1.0

# Filtering and Utils
django-cleanup>=9.0.0
annotated-types>=0.7.0

# API Documentation
drf-yasg>=1.21.8

# Development Tools
distlib>=0.3.9
filelock>=3.16.1
platformdirs>=4.3.6

# Security
django-ratelimit>=4.1.0  # For rate limiting
django-cors-headers>=4.6.0  # Already listed but relevant for security

# Dependencies
asgiref>=3.8.1
cachetools>=5.5.0
certifi>=2024.12.14
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
googleapis-common-protos>=1.66.0
httplib2>=0.22.0
idna>=3.10
inflection>=0.5.1
oauthlib>=3.2.2
packaging>=24.2
proto-plus>=1.25.0
protobuf>=5.29.3
pyasn1>=0.6.1
pyasn1-modules>=0.4.1
PyJWT>=2.10.1
pyparsing>=3.2.1
python-dotenv>=1.0.1
pytz>=2024.2
PyYAML>=6.0.2
requests>=2.32.3
requests-oauthlib>=2.0.0
rsa>=4.9
wheel>=0.45.1
requests==2.31.0
beautifulsoup4==4.12.2 

# Note: The following packages from original requirements.txt are not currently used:
# - black, flake8, isort (development tools)
# - psycopg2-binary (PostgreSQL)
# - django-storages
# - django-debug-toolbar
# - pytest, pytest-django
# - django-ratelimit
# - django-ses
# - django-redis
# - celery
# - redis